fix: update and delete authorization

+++ internal/dependencies/memdb/wh.go
@@ -50,8 +50,21 @@ func createNewWhMemDb() (*memdb.MemDB, error) {
 	return memdb.NewMemDB(schema)
 }
 
-func (s *WhDbService) Retrieve(ctx context.Context, t domain.WhType, whId string, users []string, sharedUsers []string) (*domain.Wh, *domain.DbError) {
-	txn := s.Db.Txn(false)
+func (s *WhDbService) Retrieve(ctx context.Context, t domain.WhType, whId string, userIds []string, sharedUserIds []string) (*domain.Wh, *domain.DbError) {
+	wh, dbErr := getOne(s.Db, t, whId)
+	if dbErr != nil {
+		return nil, dbErr
+	}
+
+	if slices.Contains(userIds, wh.OwnerId) || slices.Contains(sharedUserIds, wh.OwnerId) && wh.IsShared() {
+		return wh, nil
+	} else {
+		return nil, &domain.DbError{Type: domain.DbNotFoundError, Err: errors.New("wh not found")}
+	}
+}
+
+func getOne(db *memdb.MemDB, t domain.WhType, whId string) (*domain.Wh, *domain.DbError) {
+	txn := db.Txn(false)
 	whRaw, err := txn.First(string(t), "id", whId)
 	if err != nil {
 		return nil, &domain.DbError{Type: domain.DbInternalError, Err: err}
@@ -65,20 +78,23 @@ func (s *WhDbService) Retrieve(ctx context.Context, t domain.WhType, whId string
 	if !ok {
 		return nil, &domain.DbError{Type: domain.DbInternalError, Err: fmt.Errorf("could not populate wh from raw %v", whRaw)}
 	}
-
-	if slices.Contains(users, wh.OwnerId) || slices.Contains(sharedUsers, wh.OwnerId) && wh.IsShared() {
-		return wh.Copy(), nil
-	}
-
-	return nil, &domain.DbError{Type: domain.DbNotFoundError, Err: errors.New("wh not found")}
-
+	return wh.Copy(), nil
 }
 
 func (s *WhDbService) Create(ctx context.Context, t domain.WhType, w *domain.Wh) (*domain.Wh, *domain.DbError) {
 	return upsertWh(s.Db, t, w)
 }
 
-func (s *WhDbService) Update(ctx context.Context, t domain.WhType, w *domain.Wh) (*domain.Wh, *domain.DbError) {
+func (s *WhDbService) Update(ctx context.Context, t domain.WhType, w *domain.Wh, userId string) (*domain.Wh, *domain.DbError) {
+	wh, dbErr := getOne(s.Db, t, w.Id)
+	if dbErr != nil {
+		return nil, dbErr
+	}
+
+	if wh.OwnerId != userId {
+		return nil, &domain.DbError{Type: domain.DbNotFoundError, Err: errors.New("invalid owner id")}
+	}
+
 	return upsertWh(s.Db, t, w)
 }
 
@@ -93,7 +109,21 @@ func upsertWh(db *memdb.MemDB, t domain.WhType, w *domain.Wh) (*domain.Wh, *doma
 	return w.Copy(), nil
 }
 
-func (s *WhDbService) Delete(ctx context.Context, t domain.WhType, whId string) *domain.DbError {
+func (s *WhDbService) Delete(ctx context.Context, t domain.WhType, whId string, userId string) *domain.DbError {
+	wh, dbErr := getOne(s.Db, t, whId)
+	if dbErr != nil {
+		switch dbErr.Type {
+		case domain.DbNotFoundError:
+			return nil
+		default:
+			return dbErr
+		}
+	}
+
+	if wh.OwnerId != userId {
+		return nil
+	}
+
 	txn := s.Db.Txn(true)
 	defer txn.Abort()
 	if _, err := txn.DeleteAll(string(t), "id", whId); err != nil {
